WIP: Keep an array with source files for each app
Instead of one source file, allow setting multiple source files so that
we can have the desktop file, appdata file and appstream file all marked
as a source for one app.

This is supposed to fix https://bugzilla.gnome.org/show_bug.cgi?id=775814
but something is still not right there and this needs a bit more
investigation.